[split] a number of finagle bugfixes and improvements.
  * finagle/../LeastQueuedStrategy.scala: make sure we account for
  service creation failure. otherwise downed hosts will have a pegged
  queue size from the view of the strategy. this prevented proper
  rebalancing in scenarios like rolling restarts.

  * finagle/../ClientBuilder.scala: reintroduce failure accrual into the
  stack, with reasonable defaults.

  * finagle/../ServerBuilder.scala: promote buffer size parameterization
  to the child channels -- they make sense only there.

  * finagle/../ChannelService.scala: keep track of failed & cancelled
  connects as well.

  * finagle/../LoadBalancedFactory.scala: add availability as a gauge

  * finagle/../FailureAccrualFactory.scala: also account for service
  creation failures.

+++ finagle-ostrich4/src/main/scala/com/twitter/finagle/stats/OstrichStatsReceiver.scala
@@ -4,7 +4,7 @@ import com.twitter.ostrich.stats.Stats
 
 class OstrichStatsReceiver extends StatsReceiverWithCumulativeGauges {
   protected[this] def registerGauge(name: Seq[String], f: => Float) {
-    Stats.addGauge(variableName(name)) { f }
+    Stats.addGauge(variableName(name)) { f.toDouble }
   }
 
   protected[this] def deregisterGauge(name: Seq[String]) {
